requirements.txt
# Image classification with PyTorch
torch
torchvision
numpy
matplotlib
timm

# web scraping and automation
selenium

# message generation
google-genai